ProjectConfiguration

An object storing configuration used project-wide

See also

Super class

esqlabsR::ProjectConfiguration -> ProjectConfigurationRF

Active bindings

addOns

list with non default configurations

Methods

Public methods

ProjectConfigurationRF$new()

ProjectConfigurationRF$print()

ProjectConfigurationRF$addAddOnFileToConfiguration()

ProjectConfigurationRF$addAddOnFolderToConfiguration()

ProjectConfigurationRF$clone()

Inherited methods

esqlabsR::ProjectConfiguration$save()

Method new()

Initializes the ProjectConfiguration object with a specified configuration file path.

Adds new line to configuration xlsx

Read configuration from file Initialize

Usage

ProjectConfigurationRF$new(projectConfigurationFilePath = character())

Arguments

projectConfigurationFilePath

A string representing the path to the project configuration file. Print

Method print()

print prints a summary of the Project Configuration.

Usage

ProjectConfigurationRF$print(className = TRUE)

Arguments

className

Whether to print the name of the class at the beginning. default to TRUE.

Method addAddOnFileToConfiguration()

Adds an add-on file to the project configuration.

Usage

ProjectConfigurationRF$addAddOnFileToConfiguration(
  property,
  value,
  description,
  templatePath
)

Arguments

property

A string representing the name of the property to add.

value

A string representing the basename of the file to add.

description

A string providing a description of the property.

templatePath

A string representing the path of the file to add.

Method addAddOnFolderToConfiguration()

Adds an add-on file to the project configuration.

Usage

ProjectConfigurationRF$addAddOnFolderToConfiguration(
  property,
  value,
  description
)

Arguments

property

A string representing the name of the property to add.

value

A string representing the path of the value to add.

description

A string providing a description of the property.

templatePath

A string representing the path of the template file.

Method clone()

The objects of this class are cloneable with this method.

Usage

ProjectConfigurationRF$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.